+++ lib/amnesie/runner.rb
@@ -2,6 +2,7 @@ require_relative 'options'
 require_relative 'mac'
 require_relative 'process'
 require_relative 'network'
+require_relative 'persist'
 
 module Amnesie
   class Runner
@@ -10,6 +11,7 @@ module Amnesie
       @options = Options.new(argv)
       @network = false
     end
+
     def run
       if @options.mac then
         if not @network
@@ -25,6 +27,21 @@ module Amnesie
         card.apply
         puts card.to_s
         process.restart
+      end
+      if @options.persist then
+        if not @network
+          @network = Amnesie::Network.new(@options.netcard)
+        end
+        puts @network.card
+        persist = Amnesie::Persist.new(@network.card)
+        if ! persist.mac_exist?
+          puts "Create service..."
+          persist.services
+        elsif persist.mac_exist?
+          persist.update_mac
+        end
+        persist.menu_mac
+        puts persist.to_s
       end
     end
   end